project_dir/farmStack_dir/farmStack/backend/requirements.txt
# fastapi==0.109.0
fastapi
# uvicorn==0.24.0.post1
uvicorn
# motor==3.3.1
motor
# python-dotenv==0.19.0
python-dotenv